refactor: move upm-config types to own module
This will be  its own module so that it can also contain helpers related to the type which do not involve io
